Add support for Emacs customization.
Replaced several defvar forms with appropriate defcustom forms.
This change has no impact on ability to configure the ffip in Emacs initialization file as described in the home page.
It does, however, add the ability to use Emacs Easy Customization as described in Emacs manual.
Nothing is lost, something is gained for people that want to use Emacs Easy Customization.